fix(Chart): correct dependencies, move stories
