+++ libs/ui/chart/package.json
@@ -22,10 +22,14 @@
     "@amcharts/amcharts4-geodata": "^4.1.14",
     "@angular/common": "^9.1.0",
     "@angular/core": "^9.1.0",
+    "@angular/forms": "^9.1.0",
     "@angular/platform-browser": "^9.1.0",
     "@terminus/design-tokens": "^3.1.0",
     "@terminus/fe-utilities": "^1.0.0",
+    "@terminus/ui-spacing": "^1.1.3",
     "@terminus/ui-utilities": "^1.0.9",
+    "date-fns": "^2.16.1",
+    "rxjs": "^6.5.0",
     "tslib": "^1.10.0"
   },
   "scripts": {
