+++ libs/ui/chart/schematics/ng-add/index.ts
@@ -13,17 +13,17 @@ export const ngAdd = () => (tree: Tree, context: SchematicContext): Tree => {
   [
     '@amcharts/amcharts4: ^4.9.20',
     '@amcharts/amcharts4-geodata: ^4.1.14',
-    '@angular/cdk: ^9.2.4',
     '@angular/common: ^9.1.0',
     '@angular/core: ^9.1.0',
-    '@angular/flex-layout: ~9.0.0-beta.29',
     '@angular/forms: ^9.1.0',
     '@angular/platform-browser: ^9.1.0',
     '@terminus/design-tokens: ^3.1.0',
     '@terminus/fe-utilities: ^1.0.0',
+    '@terminus/ui-spacing: ^1.1.3',
     '@terminus/ui-utilities: ^1.0.9',
     '@terminus/ui-chart: ^1.0.13',
     'date-fns: ^2.14.0',
+    'rxjs: ^6.5.0',
   ].map(p => {
     const individualPackage = p.split(':');
     const nodeDependency: NodeDependency = {
