+++ stories/ui-chip/.storybook/main.js
@@ -1,8 +1,6 @@
 const addonConfigs = require('../../../../.storybook/utilities/addon-configs');
-const globalConfig = require('../../../../stories/ui/storybook/.storybook/main');
 
 module.exports = {
-  ...globalConfig,
   addons: [
     ...addonConfigs.addonsNoEssentials,
     ...addonConfigs.essentialsNoDocs,
