+++ stories/ui-chip/.storybook/preview-head.html
@@ -5,7 +5,7 @@
 
 <style type="text/css">
   html {
-    font-family: var(--ts-typography-compound-body);
+    font: var(--ts-typography-compound-body);
   }
   .sb-show-main {
     box-sizing: border-box;
